refactor: Nav bar props로 list 넘기는 방식으로 변경
